refactor: consistently use `binaryTarget` term
Right now we use the term "platform" in different contexts to signify
different things, and use either "platform" or "binaryTarget" in the
same context to mean the same thing. This commit consolidates the usage
of the latter to just "binaryTarget" as it matches the user-facing API
in the form of the `binaryTargets` property in Prisma schema. Since it
is less ambiguous, it is easy to search and replace it later if we
decide on a different name.

Closes: https://github.com/prisma/team-orm/issues/366